requirements.txt
envparse
click
python-gitlab
plyer